requirements.txt
setuptools
requests
validators
multipart
progress
crontab
colour

websocket-client
fire
jedi
simplejson

Flask
eventlet
flask-cors
flask-sock
PyJWT

hackerforms